+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-07-04
+
 ### Added
 
 - Add config option to reduce remaining time by a percentage of total growing / respawning time when fertilizing, rather instantaneously bypass the whole remaining time.
@@ -26,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial version.
 
-[unreleased]: https://github.com/nbusseneau/InstantFertilizer/compare/0.1.1...HEAD
+[unreleased]: https://github.com/nbusseneau/InstantFertilizer/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/nbusseneau/InstantFertilizer/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/nbusseneau/InstantFertilizer/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/nbusseneau/InstantFertilizer/compare/633f841a6cf39f9b82f2a85adf58882f18d461ea...0.1.0